refactor: cretae TComponents to group components types
