--- src/types/TButton/index.ts
@@ -1,9 +1,0 @@
-import { TStatusRegistration } from "../TStatus";
-
-export type TButton =  {
-    children: React.ReactNode;
-    onClick?: () => void;
-    type?: 'button' | 'submit' | 'reset';
-    variant?: TStatusRegistration;
-  }
-
